backend/worker_service/requirements.txt
celery==5.3.6
redis==5.0.3
pillow==10.2.0
av>=12,<13           # use wheel; avoid Cython build hassle
ffmpeg-python==0.2.0
python-dotenv==1.0.1
